Don't update rate_limit every time we call ossl_prov_is_running()
Thread santizer was complaining about the unlocked update to the static
variable rate_limit in the FIPS module. In practice this is only important
in the event of the module having gone into the error state. Technically
this access should be locked, but it is only used to limit the rate of
errors that are raised on the error stack. It's not particularly important
if it gets this rate limiting wrong and a lock for this seems over the top.
So we just fix the concurrent access in the normal running state in order
to appease thread sanitizer, and ignore the concurrent access in the error
state.