MAC: Undo the passing of an OSSL_PARAM array to the provider side init
The effect of having the OSSL_PARAM array passed to OSSL_FUNC_mac_init()
is that all providers must add a call to their set_ctx_params() function
if they handle params at all.  That call can as well be made in one spot
in libcrypto so that providers don't have to.

One small burden less on provider authors.