Remove internal libctx from legacy provider.
The code was originally copied from the fips provider which actually
needs a internal libctx. The libctx is now just set to the core libctx
(like the default provider).
This fixes the issue in PR #14326m, which added the first algorithm
(KDF) that fetches another algorithm (digest) - which fell over since the internal
libctx in the legacy provider was empty.